% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_boxplot.R
\name{drg_boxplot}
\alias{drg_boxplot}
\title{Create DRG Boxplot}
\usage{
drg_boxplot(data, colOpt)
}
\arguments{
\item{data}{Data frame containing data from data.cms.gov. The data can be loaded from the package
using the command "data(DRG_data)"}

\item{colOpt}{Desired column for calculation a a string: "Average.Medicare.Payments", "Average.Covered.Charges",
or "Average.Total.Payments"}
}
\value{
A ggplot object of a boxplot of medical procedure payments by DRG code
}
\description{
This function takes a data frame containing Average Medicare Payments, Average Total Payments, and Average Covered
Charges for various hospital charges organized by DRG medical procedure codes, and it creates and returns a plot of
boxplots for each DRG code for one of the data columns.
}
\examples{
~ Load Data ~
data(DRG_data)

drg_boxplot(data = DRG_data,
            colOpt = "Average.Covered.Charges")

}
