% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_stats_medicare.R
\name{drg_stats_medicare}
\alias{drg_stats_medicare}
\title{DRG Code Medicare Statistics}
\usage{
drg_stats_medicare(data, stat)
}
\arguments{
\item{data}{Data frame containing data from data.cms.gov}

\item{stat}{Desired calculation statistic as a string: "mean", "median", or "sd"}
}
\value{
A data frame containing the statistics by DRG code for Average Medicare Payments
}
\description{
This function takes a data frame containing Average Medicare Payments, Average Total Payments, and Average Covered
Charges for various hospital charges organized by DRG medical procedure codes, and it calculates the mean,
standard deviation, or median for the Average Medicare Payments by DRG code
}
\examples{
data(drg_data)

mean_medicare <- drg_stats(data = drg_data,
                                stat = "mean")

}
