% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_boxplot.R
\name{drg_boxplot}
\alias{drg_boxplot}
\title{Create DRG Plot}
\usage{
drg_boxplot(data, option)
}
\arguments{
\item{data}{Data frame containing data from data.cms.gov}

\item{option}{Desired column for calculation a a string: "Average.Medicare.Payments", "Average.Covered.Charges", or "Average.Total.Payments"}
}
\value{
A ggplot object of a boxplot of payments by DRG code
}
\description{
This function takes a data frame containing Average Medicare Payments, Average Total Payments, and Average Covered
Charges for various hospital charges organized by DRG medical procedure codes, and it creates and returns a plot of
boxplots for each DRG code for one of the data columns.
}
\examples{
data(drg_data)

drg_plot(drg_data, "Average.Covered.Charges")

}
