% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_stats_medicare.R
\name{drg_stats_medicare}
\alias{drg_stats_medicare}
\title{DRG Code Medicare Statistics}
\usage{
drg_stats_medicare(data, stat)
}
\arguments{
\item{data}{Data frame containing data from data.cms.gov}

\item{stat}{Desired calculation statistic as a string: "mean", "median", or "sd"}
}
\value{
A data frame containing the statistics by DRG code for Average Medicare Payments
}
\description{
This function takes a data frame containing Average Medicare Payments for different medical procedures and their
DRG codes at various hospitals, and it calculates a either the standard deviation, mean, or median for each DRG
code and returns a data frame for further manipulation.
}
\examples{
data(DRG_data)

mean_medicare <- drg_stats_medicare(data = DRG_data,
                                stat = "mean")

}
